# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizer/clients:
    get:
      operationId: listOrganizerClients
      tags: [organizer]
      summary: List the signed-in organizer's clients with budget snapshots
      responses:
        "200":
          description: Clients with a quick-glance budget snapshot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizerClientSummary"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an organizer account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrganizerClient
      tags: [organizer]
      summary: Add a client, optionally linking a couple already on Bazm by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizerClientInput"
      responses:
        "201":
          description: The created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerClientSummary"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an organizer account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizer/clients/{id}:
    get:
      operationId: getOrganizerClient
      tags: [organizer]
      summary: One client with their linked plan, chosen tiers, and pending decisions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerClientDetail"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an organizer account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizer/summary:
    get:
      operationId: getOrganizerSummary
      tags: [organizer]
      summary: Weekly outreach summary — upcoming weddings and unresolved clients
      responses:
        "200":
          description: Summary counts and upcoming weddings in the next 30 days
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerSummary"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an organizer account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  # ---------------------------------------------------------------------------
  # Seller products (role=vendor, scoped to the owned vendor)
  # ---------------------------------------------------------------------------
...
components:
  schemas:
...
    PlannerCategory:
      type: string
      enum: [venue, catering, decor, music, photography, clothing]
    TierLevel:
      type: string
      enum: [budget, standard, premium]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OrganizerClientSummary:
      type: object
      required:
        [id, clientName, status, linked, hasPlan, pendingCount, decidedCount]
      properties:
        id:
          type: integer
        clientName:
          type: string
        weddingDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        linked:
          type: boolean
          description: True when the client is linked to a couple account on Bazm.
        hasPlan:
          type: boolean
        budgetTotal:
          type: ["integer", "null"]
          description: Sum of estimated category costs in the linked plan (UZS).
        undecidedTotal:
          type: ["integer", "null"]
          description: Estimated cost tied to categories with no chosen vendor tier.
        pendingCount:
          type: integer
        decidedCount:
          type: integer
        daysUntilWedding:
          type: ["integer", "null"]
    CreateOrganizerClientInput:
      type: object
      required: [clientName]
      properties:
        clientName:
          type: string
          minLength: 1
          maxLength: 120
        weddingDate:
          type: ["string", "null"]
          description: ISO date (YYYY-MM-DD), or null if not set yet.
        notes:
          type: ["string", "null"]
          maxLength: 2000
        shareToken:
          type: ["string", "null"]
          maxLength: 200
          description: >
            Optional budget-plan share token the couple gave the organizer
            (the token from their read-only share link). When it matches a
            plan, the client is linked to that couple so their budget plan
            appears. Holding the token is the couple's consent to share.
    OrganizerPlanSummary:
      type: object
      required: [city, guestCount, estimatedTotal, undecidedTotal, createdAt]
      properties:
        city:
          type: string
        guestCount:
          type: integer
        totalBudget:
          type: ["integer", "null"]
        estimatedTotal:
          type: integer
        undecidedTotal:
          type: integer
        createdAt:
          type: string
          format: date-time
    OrganizerChosenTier:
      type: object
      required: [category, estimatedCost]
      properties:
        category:
          $ref: "#/components/schemas/PlannerCategory"
        tierLevel:
          oneOf:
            - $ref: "#/components/schemas/TierLevel"
            - type: "null"
        tierName:
          type: ["string", "null"]
        vendorName:
          type: ["string", "null"]
        estimatedCost:
          type: integer
    OrganizerPendingCategory:
      type: object
      required: [category, estimatedCost]
      properties:
        category:
          $ref: "#/components/schemas/PlannerCategory"
        estimatedCost:
          type: integer
    OrganizerClientDetail:
      type: object
      required: [id, clientName, status, linked, chosenTiers, pendingCategories]
      properties:
        id:
          type: integer
        clientName:
          type: string
        weddingDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        linked:
          type: boolean
        daysUntilWedding:
          type: ["integer", "null"]
        plan:
          oneOf:
            - $ref: "#/components/schemas/OrganizerPlanSummary"
            - type: "null"
        chosenTiers:
          type: array
          items:
            $ref: "#/components/schemas/OrganizerChosenTier"
        pendingCategories:
          type: array
          items:
            $ref: "#/components/schemas/OrganizerPendingCategory"
    OrganizerUpcomingWedding:
      type: object
      required:
        [id, clientName, weddingDate, daysUntilWedding, status, pendingCount]
      properties:
        id:
          type: integer
        clientName:
          type: string
        weddingDate:
          type: string
        daysUntilWedding:
          type: integer
        status:
          type: string
        pendingCount:
          type: integer
    OrganizerSummary:
      type: object
      required: [totalClients, upcomingCount, unresolvedClientCount, upcoming]
      properties:
        totalClients:
          type: integer
        upcomingCount:
          type: integer
        unresolvedClientCount:
          type: integer
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/OrganizerUpcomingWedding"